requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyactiveresource==2.2.2
PyJWT==2.3.0
PyYAML==6.0
ShopifyAPI==11.0.0
six==1.15.0
Werkzeug==2.0.2
